--- src/main/java/app/entities/RoofType.java
@@ -1,6 +1,0 @@
-package app.entities;
-
-public enum RoofType {
-    FLAT,
-    RAISED
-}
