+++ src/main/java/app/services/StructureCalculationEngine/CalculationStrategy.java
@@ -1,4 +1,7 @@
-package app.entities;
+package app.services.StructureCalculationEngine;
+
+import app.services.StructureCalculationEngine.Entities.PlacedMaterial;
+import app.services.StructureCalculationEngine.Entities.Structure;
 
 import java.sql.SQLException;
 import java.util.List;
