+++ src/main/java/app/services/StructureCalculationEngine/RoofType.java
@@ -1,0 +1,6 @@
+package app.services.StructureCalculationEngine;
+
+public enum RoofType {
+    FLAT,
+    RAISED
+}
